# BackEnd/docs/swagger.yaml
basePath: /
definitions:
...
  web_doscom_internal_database_model.RegisterRequest:
    properties:
      email:
        type: string
      fullname:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - fullname
    type: object
  web_doscom_internal_database_model.UserPatch:
    properties:
      email:
        type: string
      fullname:
        type: string
      username:
        type: string
    type: object
  web_doscom_internal_database_model.UserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: API Documentation Web Doscom
  title: Web Doscom API
  version: "1.0"
paths:
  /api/v1/admin/:
    get:
      description: Mendapatkan daftar semua user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web_doscom_internal_database_model.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Super_Admin
    post:
      consumes:
      - application/json
      description: Membuat user baru
      parameters:
      - description: Email, FullName, Role
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/web_doscom_internal_database_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create new user
      tags:
      - Super_Admin
  /api/v1/admin/{id}:
    get:
      description: Mendapatkan user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Super_Admin
    put:
      consumes:
      - application/json
      description: Mengupdate data user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/web_doscom_internal_database_model.UserPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Super_Admin
  /api/v1/admin/superadmin:
    post:
      consumes:
      - application/json
      description: Membuat user dengan role superadmin (hanya untuk admin/superadmin)
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/web_doscom_internal_database_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create superadmin user
      tags:
      - Super_Admin
...
  /api/v1/koor/:
    get:
      description: Mendapatkan daftar semua user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web_doscom_internal_database_model.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Koor
    post:
      consumes:
      - application/json
      description: Membuat user baru
      parameters:
      - description: Email, FullName
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/web_doscom_internal_database_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create new user
      tags:
      - Koor
  /api/v1/koor/{id}:
    get:
      description: Mendapatkan user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Koor
    put:
      consumes:
      - application/json
      description: Mengupdate data user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/web_doscom_internal_database_model.UserPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Koor
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"